+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unrealeased]
 
+### Fixed
+
+- Fix course enrollment count shouldn't include the hidden runs.
+
 ## [2.28.1]
 
 ### Fixed